refactor: rename tools/core/ to tools/shared/ and remove barrel re-exports
- Rename src/tools/core/ → src/tools/shared/ for a clearer name that
  reflects the actual intent (mode-agnostic logic shared across variants)
- Remove re-export shims from actor.ts that proxied core/ symbols
- Update tools/index.ts to import directly from shared/ instead of
  routing through actor.ts
- Update all import sites: server.ts, actor.ts, index.ts, and unit test
